+++ src/sim/csimulation.cc
@@ -832,8 +832,8 @@ class StaticEnv : public cEnvir
     virtual unsigned long getUniqueNumber() override  { unsupported(); return 0; }
     virtual bool idle() override  { return false; }
     virtual void attachDebugger() override {}
-    virtual void getImageSize(const char *imageName, int& outWidth, int& outHeight) override {unsupported();}
-    virtual void getTextExtent(const cFigure::Font& font, const char *text, int& outWidth, int& outHeight, int& outAscent) override {unsupported();}
+    virtual void getImageSize(const char *imageName, double& outWidth, double& outHeight) override {unsupported();}
+    virtual void getTextExtent(const cFigure::Font& font, const char *text, double& outWidth, double& outHeight, double& outAscent) override {unsupported();}
     virtual double getAnimationTime() const override {return 0;}
     virtual double getAnimationSpeed() const override {return 0;}
     virtual double getRemainingAnimationHoldTime() const override {return 0;}
